gpg-webservice-rest/requirements.txt
flask
flask-sqlalchemy
flask-cors
werkzeug
cryptography
argon2-cffi
pytest
python-dotenv
